Add files via upload
This paper studies the integration of “Artificial Intelligence (AI)” and the “Internet of Things
(IoT)” for predicting energy consumption in “Smart Homes”. With the widespread acceptance
of “Smart Home Technologies”, as well as the ever-growing need for energy efficiency, it is
essential to design intelligent systems that optimize energy usage. The data set in this research
had the following information relating to appliance usage, temperature set, status of occupants,
and so forth. Applying machine learning methods for energy prediction, a Random Forest
Regressor model is implemented on such a data set to outline significant factors contributing to
consumption and predict patterns in subsequent consumption. The outcome reveals that the
duration of usage of appliances and the temperature setting are the most important factors
determining energy consumption, whereas other variables like day of the week and seasonal
variations also come into play.